Enable VisualEditor by default in NS_USER as well as $wgContentNamespaces
Initially we were only deploying to $wgContentNamespaces, but it's highly likely
that users will want to edit their own user pages in VE anyway. Note that this is
how it's been deployed for Wikimedia for a year now.

Additionally, this helps fix a bug in the browser tests that assumed NS_USER was
a VE-able namespace when that was only true in production config, and not for new
wikis (such as created during automated setup).

Bug: 57104
Change-Id: Icebc24f82fb49e9e80388b6202efa366360ce33c