[browser test] refactor bullets test to use API etc.
Get rid of all hard-coded values for both the target page
and also the text on the target page, which is now all set
in the .feature file.

Note that we're deconstructing the @make_selectable hook
by doing this. @make_selectable may become obsolete when
all the tests are refactored to use APIPage

Change-Id: Iad1b160d9d0274effb8504f276bd4816932d299c