[BrowserTest] Language Screenshot (browser test and their screenshots) for Headings and Formatting
This is an illustration of how one can get a screenshot of
the Headings pull-down menu and formatting pull-down menu for the Visual Editor User guide[1].
To see how it works:
* Set environment variable LANGUAGE_SCREENSHOT_CODE to which ever
language you are planning to run browser test for.
(Eg: de for German/en for English)
* Set environment variable LANGUAGE_SCREENSHOT_PATH (optional)
to store the screenshot in that path. If not provided the ./screenshots folder will be used.
* Run the browser test as usual using cucumber or you can use @language_screenshot tag while executing
(bundle exec cucumber --tags @language_screenshot)

[1]: https://www.mediawiki.org/wiki/Help:VisualEditor/User_guide

Bug: 62737
Change-Id: I0f27aec482458b6f04aa913426042f78695d451d