Followup 54458c8: make newly RTFified icons actually work everywhere
Right now our system is so horrible that we need three separate
CSS rules to make the icons appear in three separate places:

1. Toolbar (with contextual directionality)
2. Context menu
3. Top-left corner of dialogs

Also, the icons in the context menu should use contextual directionality
but are using the document directionality because the context doesn't
compute and set a class for the context directionality. This is only
done by the toolbar. It used to work because the context menu *was*
a toolbar, but that changed recently.

Change-Id: I3ebc1ba7dfcd384a2a62664e55a4a77c8dedd513