ve.ce.MWExtensionNode: Don't escape content of wikitext tags on preview
XML-like tags in wikitext are not actually XML and don't expect their
contents to be escaped. The contents are usually not parsed as
HTML/XML anyway.

It means that it is not possible for a tag '<foo>…</foo>' to contain
the string '</foo>', but that is prevented by change Ia566452a.

Bug: 57429
Change-Id: I2516a897c85b4a3b79a34ff6cad5d124fbc30170