MWTemplatePage: Account for template.title being null
Example:
* Transclusion with html comment in target
  {{ Unsigned <!-- ignored -->  |Example}}.
* Transclusion with wikitext expression as target
  {{ {{echo|Unsigned}} <!-- ignored -->  |Example}}.

The first case was handled already (Parsoid strips the comment
when it normalises the target title). The second case, however,
is not handled and ends in template.title being null.

This was already documented (MWTemplateModel#getTitle returns
{string|null}), but not used correctly.

The title of the dialog is unaffected as it fallsback to
displaying target.wt instead of target.title.

Bug: 66724
Change-Id: Ib6696ac3538f6cb9e93da2f6f8666f373247eeb4