Change label in dialog: "Change image" -> "Choose a different image"
The media dialog has an "Apply changes" primary action and a "Change image"
secondary action. These two labels are confusingly similar. "Choose a
different image" is a long label, but even on mobile a longer label
should fit nicely in this context, and this seems to me to avoid the
issue of ambiguity.

Change-Id: Id8be8b4cf70f7eee689b72d88b0aeac20c2eca93