"Fix" hack in ve.test.utils to prevent immediate surface setup
This used to work because loading i18n files took time. However,
we now don't register any message paths with the platform, so
platform initialization is now synchronous, which breaks the hack.

Instead, explicitly assign an unresolved deferred to the platform's
initialization promise, and resolve it after the mw.Target has been
constructed.

*barf*

Change-Id: I287cb86ce8c4c534fd723532f34463359a8ccc7c